// prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Genre: {
      readonly id: CodecTypes["pg/text@1"]["output"];
      readonly name: CodecTypes["pg/text@1"]["output"];
    };
    readonly Movie: {
      readonly createdAt: CodecTypes["pg/timestamptz-temporal@1"]["output"];
      readonly description: CodecTypes["pg/text@1"]["output"] | null;
      readonly id: CodecTypes["pg/text@1"]["output"];
      readonly releaseYear: CodecTypes["pg/int4@1"]["output"] | null;
      readonly title: CodecTypes["pg/text@1"]["output"];
      readonly updatedAt: CodecTypes["pg/timestamptz-temporal@1"]["output"];
    };
    readonly MovieGenre: {
      readonly genreId: CodecTypes["pg/text@1"]["output"];
      readonly movieId: CodecTypes["pg/text@1"]["output"];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Genre: {
      readonly id: CodecTypes["pg/text@1"]["input"];
      readonly name: CodecTypes["pg/text@1"]["input"];
    };
    readonly Movie: {
      readonly createdAt: CodecTypes["pg/timestamptz-temporal@1"]["input"];
      readonly description: CodecTypes["pg/text@1"]["input"] | null;
      readonly id: CodecTypes["pg/text@1"]["input"];
      readonly releaseYear: CodecTypes["pg/int4@1"]["input"] | null;
      readonly title: CodecTypes["pg/text@1"]["input"];
      readonly updatedAt: CodecTypes["pg/timestamptz-temporal@1"]["input"];
    };
    readonly MovieGenre: {
      readonly genreId: CodecTypes["pg/text@1"]["input"];
      readonly movieId: CodecTypes["pg/text@1"]["input"];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly genres: {
      readonly id: CodecTypes["pg/text@1"]["output"];
      readonly name: CodecTypes["pg/text@1"]["output"];
    };
    readonly movie_genres: {
      readonly genreId: CodecTypes["pg/text@1"]["output"];
      readonly movieId: CodecTypes["pg/text@1"]["output"];
    };
    readonly movies: {
      readonly createdAt: CodecTypes["pg/timestamptz-temporal@1"]["output"];
      readonly description: CodecTypes["pg/text@1"]["output"] | null;
      readonly id: CodecTypes["pg/text@1"]["output"];
      readonly releaseYear: CodecTypes["pg/int4@1"]["output"] | null;
      readonly title: CodecTypes["pg/text@1"]["output"];
      readonly updatedAt: CodecTypes["pg/timestamptz-temporal@1"]["output"];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly genres: {
      readonly id: CodecTypes["pg/text@1"]["input"];
      readonly name: CodecTypes["pg/text@1"]["input"];
    };
    readonly movie_genres: {
      readonly genreId: CodecTypes["pg/text@1"]["input"];
      readonly movieId: CodecTypes["pg/text@1"]["input"];
    };
    readonly movies: {
      readonly createdAt: CodecTypes["pg/timestamptz-temporal@1"]["input"];
      readonly description: CodecTypes["pg/text@1"]["input"] | null;
      readonly id: CodecTypes["pg/text@1"]["input"];
      readonly releaseYear: CodecTypes["pg/int4@1"]["input"] | null;
      readonly title: CodecTypes["pg/text@1"]["input"];
      readonly updatedAt: CodecTypes["pg/timestamptz-temporal@1"]["input"];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;